requirements.txt
accelerate==0.24.1
addict==2.4.0
aiobotocore==2.7.0
aiofiles==23.2.1
art==6.1
bitsandbytes==0.41.3.post2
colorama==0.4.6
einops==0.8.0
evaluate==0.4.0
fire==0.5.0
fschat==0.2.34
gcsfs==2023.10.0
gradio
huggingface_hub==0.20.1
mamba_ssm==1.2.0.post1
numba==0.58.1
numpy==1.24.4
openai==1.23.6
optimum==1.13.2
pandas==2.0.3
peft==0.6.0
pynvml==11.5.0
pytest==8.2.0
PyYAML==6.0.1
Requests==2.31.0
s3fs==2023.10.0
safetensors==0.4.3
setuptools==52.0.0.post20210125
termcolor==2.4.0
tqdm==4.66.1
transformers==4.37.0
wandb==0.16.1
xformers==0.0.23.post1+cu118
--extra-index-url https://huggingface.github.io/autogptq-index/whl/cu118/
auto-gptq==0.5.1
packaging
tokenizers==0.15.0
bitsandbytes>=0.41.1
deepspeed
PyYAML>=6.0
datasets>=2.15.0
flash-attn==2.3.3
sentencepiece
hf_transfer
numpy>=1.24.4
bert-score==0.3.13
rouge-score==0.1.2
scipy
scikit-learn==1.2.2
tensorboard
flash_attn